Start of the contextual menu implemmentation
Areas of the original code were commented out, and I implemented the contextual menu and started to make some features work.
Share sheet and suggest other author already implemented.
Add to Favorites and Add to Folder are still do be implemented. Need a little bit of help here.
Changed the radius of the button to 14, which is the same as the contextual menu effect.
Changed the sort icon.