Working on a flexible grid, but not having much luck as of yet
I'm testing different cell heights and how it affects the View. The issue I'm trying to fix now is making the cells flow with the correct spacing between them, and not stay in a fixed position in the middle of the allocated space. After that, I'll try to use GeometryReader to see what height the cell should be related to the amount of text presented.